Add triggers to execute functions on changes.
This patch adds triggers to the resource manager. The triggers are added to a
resource and when the value changes, the triggers are executed. Once we have a
real Executor, the functions registered as triggers will be queued for
execution rather than executed in-place.

There is no support for removing triggers.