WL#6123: Basic HA Manager Framework
This patch implement beginning of the Hub.

The patch contain the core manager, a simple executor class, and a service
manager that currently just start a single XML-RPC protocol server. The
executor module also contain two function decorators ("primitive" and
"coordinated") used to implement the procedures of the system.

The service manager automatically loads services from the mysql.hub.services
package, but does not include examples of such a service.

The patch also demonstrate how to implement new commands by creating commands
for starting and stopping the command center.  You will find the implementation
of the "shutdown" command in the next patch, since it requires a service.

The patch includes configuration file handling and a sample configuration file.
This configuration file can be used with the two commands provided with this
patch.