There is one single function to execute any statement. However, if one tries to fetch a result set after executing a statement that generates none (e.g. INSERT), the python connector raises an exception.
To avoid problems, we have tried to use Cursor._have_result and
skip any call to a fetch operation. Unfortunately, the variable
is not available in all versions what is causing problems.

To fix the problem, we check the reported message in the exception
and ignore the problem if it stems from the lack of a result set.