HAM-18: Persister Management
This patch introduces management of the persister by assigning one persister
for each thread started. To set up the persister subsystem, persitence.init
have to be called which will create the necessary databases in the server. To
initialize the persistence system for a thread, the persistence.init_thread
function have to be called.

Several tests have been enabled and connection information for the persister
server have to be given to test.py when starting the tests. Some tests also
require servers running starting at the address 13000, so ensure that some
servers are running with these ports. The easiest approach is to use MTR to
start a bunch of scratch servers.