Post-push fix for ham-8.
There were references to localhost:13000 in the test_server_events
what could have caused failures when the server was not started.

To fix that, the patch uses MySQLInstances.get_uri(0) which returns
a valid address based on test's configuration. It also sets the
name of the Executor and Service's threads to ease debugging.