After promoting a candidate slave to a master in a failover operation, its replication threads were not stopped. So the new elected master would keep trying to contact the dead slave, cluttering the error log.
To avoid this problem and any other potential problem, we stop the
replication threads and reset the replication configuration. The
current solution though does not check if the server is part of a
hierarchical setup.

Notice this is not a problem as we do not support hierarchical
setups in the current version. In the future, we will improve this
solution.