In this patch, we fix the server's connection pool.
We had to revisit how the connection pool was implemented because when
it was designed we had assumed that there would be a pool per object.
However, after some discussions on the persistence layer, we figured
out that there should be a pool shared by all objects because a
MySQLServer can be dynamically created and released at any time.

We also took the opportunity to merge the Server into MySQLServer.
We have done this because there will be no support to other servers
in a near future and the abstracted class lost its meaning after the
removal of the connection pool.