This patch refactored the event-driven interface and introduced the concept of procedures and removed the job object from the operations's parameter.
Basically a procedure defines the context within which an operation is
executed. Explicitly, an operation is a code block to be executed and
is named a job.

Any job must belong to a procedure whereas a procedure may have several
jobs associated to it. When job is created and is about to be scheduled,
it is added to a set of scheduled jobs. Upon the end of its execution,
it is moved from the aforementioned set to a list of executed jobs. 
During the execution of a job, new jobs may be scheduled in the context
of the current procedure.

A procedure is marked as finished (i.e. complete) when its last job
finishes. Specifically, when a job finishes and there is no scheduled
job on behalf of the procedure.