HAM-80: Documentation is not being generated.
The documentation was not being generated because the python connector
was not found. To fix this, we have updated the python path so that it
could find the connector provided it was properly installed.

We also have taken the opportunity to integrate the routine that generates
the documentation within the setup.py. So after this patch, one will need
to run the following command to generate the documentation:

    python setup.py build_docs