HAM-85: Problems with --daemonize.
The option --daemonize was requiring either true or false when
it was supposed not to require any value. The Syslog Handler
used when the aforementioned option is set was not properly
initialized due to a missing import: logging.handlers.

We fixed these problems and took also the opportunity to fix
the following issue.

The Command.add_options() was generating, in some platforms, the
following error: "TypeError: 'NoneType' object is not iterable".
This was happening because the Command.command_options property
was None by default. In order to fix this issue, we have changed
it into [].