HAM-94: Mismatch between fabric and connector python
Python connector is using sharding.lookup_servers(**dictionary) to
retrieve information on servers responsible for a shard. However,
fabric provides the sharding.lookup interface and the xml-rpc does
not support keyword parameters.

In this patch, we renamed sharding.lookup to sharding.lookup_servers.
There will be another patch to fix the connector python.