HAM-61: Extend the server's properties and life-cycle.
Currently it is not possible to identify which servers have failed without
trying to access them. This happens due to the lack of a status field
associated to the server.

This patch fixes this problem by introducing the following server status:

- RUNNING - Right after adding a server to a group, it is alive and kicking,
  otherwise it would not be possible to add it to.
- OFFLINE - It should be used before shutting down a server otherwise the
  failover routine will trigger notifications. If users, however, want to
  bring a master offline they must firstly run a switchover.
- SPARE - This is used to make a slave not eligible to become a master in a
  switchover or failover operation. Nor are users redirected to it to execute
  read-only transactions.
- FAULTY - This is usually set by the failover routine and indicates that are
  a server is not reachable.

It also introduces the notion of active and inactive groups. By default a
group is inactive which means that the failover routine will not monitor
the group in order to check if any server has failed.