HAM-74: Add version checking
This patch introduces the necessary changes to check the version of the
managed servers, their uuids and whether registered users have the
necessary privileges, i.e. root.

The current version of Fabric only supports managing MySQL Server 5.6.8
or later and does allow to add the state store as a managed server.