HAM-78: Automatically configure an added server as slave.
Servers added to a group are not automatically configured, for example
as a slave, and as such cannot automatically be used to provide
high-availability. In order to make a server a slave, one has to
execute a switchover or failover operation.

This patch fixes this behavior and automatically sets a server as a
slave of the current master if there is any.