HAM-102: MySQL Fabric manage stop hangs when we interrupt in the fabric start page HAM-103: Fabric manage setup hangs when the corresponding server is not started.
This patches creates the necessary means to avoid blocking fabric
while accessing the state store. It does so by defining a connection
timeout on the state store, catching exceptions that are raised while
accessing it and retrying to access the state store thus hiding
sporadic failures.

Notice that the use of a explicit timeout it is fine in the context
of the state store. However, we need to revisit the computational
model in regarding the managed servers and check when and how a timeout
could be introduced to avoid blocking fabric.

This patch also makes sure that CTRL + C does not block fabric when
threads are not daemons. We have made all threads daemon.