HAM-63: Implement compensating operations (Part-II).
This patch introduces the necessary changes to save the Executor's queue
and restore it after a crash thus providing the ability to resume
execution after failures.

Information on a job is written to the checkpoint table upon scheduling
it to be executed. So before the execution any job has the started and
finished columns set to NULL. Upon starting executing a job, the started
column is set to now() and, after finishing the execution, the finished
column is set now().

Any job which has started is recovered and rescheduled. Jobs which
haven't started are automatically rescheduled.

This patch ensures that an event that triggers several jobs have the
jobs atomically scheduled.