HAM-52: Mismatch between service and logging.
The modules which were part of the service package were automatically
loaded wihout the "mysql.hub.services" prefix thus disabling the
possibility of using logging.getLogger(__name__).

The problem was fixed in the context of HAM-108, HAM-112 which were
approved and pushed to trunk. The patch has changed the way such
modules are loaded thus solving the problem.