HAM-109: Replication topology fails after a switchover/promote.
The problem happens because the switchover/failover/promote operation
issues a CHANGE MASTER without setting the password to an empty string
when it is supposed to be empty. Doing to this bug, the slave ends up
with a wrong password being stored in its repository and any attempt to
access the master fails and by consequence the replication topology
becomes broken.

To fix the problem, we explicitly set the password to an empty string
when the password is supposed to be empty.