HAM-136: logger.setLevel("INFO") does not work with python 2.6
Python 2.6 does not support passing a string as the logging level in
logger.setLevel() and as a consequence Fabric is not printing out any
information. Although this feature is not supported, python does not
throw any error or warning.

To fix the problem, we convert the string to an internal representation
which works with all the versions. For example, "INFO" to logging.INFO.