HAM-52: Mismatch between service and logging.
The services are loaded by services/_init_.py which ignores the current
hierarchy. For example, the module mysql/hub/services/server.py is
loaded as server instead of mysql.hub.services.server.

This causes problems with the logging infra-structure which expects
that every module has the following prefix mysql.hub.*.

To fix that, we have changed how the services modules are loaded.