HAM-173 Hardcoding of server addresses in sharding tests causing failure
Problem:

In the sharding code there are places where the address
of the servers are hard coded as localhost:13002,
localhost:13003, etc. This causes failures in the cases
where the server address does not start at localhost:13000.

Solution:

Use utility functions to start at dynamic server addresses.
The hardcoding of the server addresses in the code should
be removed.

Implementation in current patch:

The patch replaces hardcoded server addresses with the
utility methods from MySQLInstances, for e.g.
- "address":"localhost:13001", 
+ "address":MySQLInstances().get_address(0),