HAM-174 test_shard_split does not wait for the shard to sync all the updates from the global server
Problem
-------

The test test_shard_split checks if the shards have got all
the updates before replication can sync everything from the
global server.

solution
--------

Although this is not a long term solution, add a delay to allow
the shard to sync all the updates from the global server.

In this patch
-------------

A sleep is added to provide a delay in the test.