HAM-170: test_check_no_healthy_slave is sporadically failing
The test case was failing because the slave had stopped with
an error (i.e. table already exists) and we simply reset and
started it assuming that the failure would immediately happen
after the start was issued. However, this is not true as it
needs to fetch the information from the master and tries to
apply the events.

To fix the problem, we call wait_for_slave_gtid() and make sure
that the slave has failed again before checking its status.

We also took the opportunity to remove some trailing spaces from
the test case.