HAM-201: Commands should return True to indicate success instead of          False
Due to limitations in the xml-rpc, it is not possible to return None.
So in those procedures that None were returned were replacing it by
False. However, the False value conveys the wrong idea that a
procedure was not correctly executed.

To fix the problem, True is returned if the procedure successfully
executes. Otherwise, False is returned.

We also took the opportunity to fix a minor problem in the code that
checked if the caller should wait for the procedure execution.