HAM-194 Group check_group_availability is showing below error if a         server is down
The check_group_availability was printing out a stack trace if it
did not manage to access a server. Besides, it was reporting RUNNING
if the failure detector was off although the server was not reported
as alive.

We have fixed that by suppressing the stack trace and making the
output consistent.