HAM-222: Use rotating log file by default
The default is to write logging information to syslog, which both clutters the
syslog output and also uses a different format than syslog usually does.

This patch uses a rotating log file handler as default and also change the
format for indicating logging. Instead of using a syslog address it now uses a
URL. A 'file' scheme indicate a rotating log file and 'syslog' indicate syslog
output.

The patch also add the install prefixes into the configuration file,
making the information available to Fabric. This means that configuration files
are now stored under /usr/local/etc/fabric if the install prefix is /usr/local
and /etc/fabric if the install prefix is / or /usr.