HAM-183: Define a single interface to trigger either a switchover          or failover
In HA, there are two important notions: failover and switchover.

Although these are distinct routines, there is no need to make them
directly available to users. In this patch, we hide the notion of a
failover and switchover procedure behind a unified command which is
named promote and can be used as follows:

fabric group promote group_name [candidate]