HAM-190: Extending the underlying framework for RANGE sharding to allow its usage in HASH based sharding.
We need to extend the underlying framework for RANGE sharding to allow its usage in HASH based sharding.

Introduction
------------

The RANGE based implementation currently uses a upper bound and a lower bound.
A typical shard range definition would look like the following,
LB 	UB 	shard_id
0 	1000 	1
1001 	2000 	2
2001 	3000 	3
3001 	4000 	4

A consistent hashing implementation is actually a RANGE based implementation in
the backed. The only difference between the above implementation and the requi-
rement of a consistent HASH RING is that, the HASH RING builds over a circular
HASH RANGE ring.

The range based implementation used in a consistent HASH ring is built on a
circular ring of values, with ranges cut on the ring. The RANGE intervals are
cut on this circular ring. The maximum cut being followed by the minimum value
yet again.

Modification Required in the RANGE implementation
-------------------------------------------------

The RANGE implementation needs to be altered to accomodate values in a circular
HASH ring. In order to achieve this the above RANGE representation would look
like the following,
LB 	shard_id
0 	1
1001 	2
2001 	3
3001 	4

As can be seen the above lower bounds that are defined helps form a RING with
3001 circling over to 0.

The above shard table can also be used to store the cuts in a HASH ring.