HAM-245: Move shard_mapping_id from shards to shard_ranges
The shard_mapping_id in FABRIC represents a sharding index. In order to support cases where the user needs more than one shard index in a
shard group, we need to be able to map a shard_id to several shards. In consideration with the fact that we are implementing only RANGE and
HASH based sharding, the simplest solution to implement support for more than one index is to move the shard_mapping_id index into the
shard_ranges table.

Currently the mapping between the shard_id and the shard_mapping_id is maintained in the shards table. The shard_mapping_id is the
primary key in the shards table. Hence we can maintain only one mapping of a shard_mapping_id to a shard id in this table. When we need
to implement multiple mappings this schema becomes a problem.

The easiest solution is moving shard_mapping_id from shards into shard_ranges. Considering that we do not plan to implement list based
sharding in the initial version, this offers a simple solution for this problem.

New Schema
----------

shards

-------------------------------
| shard_id | group_id | state |
-------------------------------

shard_ranges

-----------------------------------------------------------
| shard_mapping_id | lower_bound | upper_bound | shard_id |
-----------------------------------------------------------