HAM-62 Define the appropriate concurrency control mechanism among        jobs - Part I
This patch creates a LockManager and Scheduler objects that will be
used to safely schedule concurrent procedures. The LockManager can
be used as follows:

(First Solution):

Caller:
    lock_manager.enqueue(procedure, objects)
    procedure.wait()

Executor:
    lock_manager.get(procedure)
    execute_procedure(procedure)

(Second Solution):

Caller:
    procedure = lock_manager.lock(procedure, objects)
    execute_procedure(procedure)

Procedures can be scheduled with high priority. In this case, any
procedure that conflict with the current request is aborted.

The Scheduler object uses the second approach. Currently the first
approach is only useful in the test cases. In the future, we will
compare which solution provides the best performance.

Note though that the current patch does not contain any code to
determine the set/list of conflicting objects. This will be
submitted as a second patch.