HAM-191 HASH based sharding.
The shard ranges table stores the shard key range information along with the mapping on which groups
the shard keys can be found. Currently the shard ranges table has the following schema

--------------------------------------------
|shard_mapping_id | lower_bound | shard_id |
--------------------------------------------

The shard_ranges table will be used by the HASH sharding implementation to store the underlying ranges
that drive the lookups on the consistent hashing ring. The datatype of the lower_bound column needs to
be altered to store the boundaries of the ranges on the hashing ring.

The ideal datatype of choice is VARBINARY.