HAM-264: manage stop throws an exception
The shutdown operation was not waiting until a reply was sent back to
the caller notifying that the request had been received and the server
was about to shutdown. So the caller thought that the server had died
reporting an xml-rpc failure.

We fixed the problem by making the server wait until all the session
threads stop thus guaranteeing that one of them processed the shutdown
request and replied back to the caller.