HAM-185: Setting a server's status to RUNNING should automatically          make it a slave
Setting a server's status to RUNNING when the server was either FAULTY
or OFFLINE was not triggering any change to configuration. This could
lead to configuration issues as the recovered server could be point
to an invalid master, for example.

To fix this problem, we reconfigure the server any time its status
is changed from FAULTY/OFFLINE to RUNNING.