HAM-269: Number of concurrent executors are not being set properly in          mysqlfabric
Number of concurrent executors are not being set properly in
mysqlfabric because while loading the commands the Executor
object which is a singleton is created with a default number
of executors thus ignoring the configuration file. So it was
not possible to set the number of executors through the
constructor when the start command was executed.

To fix this problem, we have introduce a set method.