HAM-267: There is no way to configure server and client individually          from the same config
It was not possible to bind Fabric to any ip-address and use the same
configuration file to manage the server. Basically, Fabric uses
":port" to bind the XML-RPC server to any address but a client willing
to locally perform administrative services could not use the same
configuration file.

To fix the problem, we make the client use "localhost:port" whenever
it finds ":port".