HAM-193: Stack traces are being printed out when it is not really          necessary
Stack traces were being printed out when it was not really necessary.
For example, it was printed out while trying to create a group which
already had been created.

To fix this problem, we have replaced the logger.exception() by
logger.debug(). We also took the opportunity to fix some mistakes in
the errors.py module and improve the code:

  . Removed UnknownEventError because it was not being used.
  . Removed PersistenceError because it was not being used.
  . Made ShardingError a sub-class of ServiceError.
  . UuidError was not supposed to be a sub-class of DatabaseError.
  . TimeoutError was not supposed to be a sub-class of ProgrammingError.
  . Organized the errors in the file.
  . Removed unused logging.getLogger.
  . Removed trailing spaces.
  . Improved some logging messages.