HAM-285: Error is not proper if promote a faulty status servers          in a group
If we tried to promote a faulty server to master, the following error
was printed out:

ServerError: ('Server (%s) is not either running or a spare.',
              ('ee74befe-f2b5-11e2-ac33-00144f40b86c',)

Although the content of the message was fine, it was not printed out
correctly. The correct output should be:

ServerError: ('Server (ee74befe-f2b5-11e2-ac33-00144f40b86c) is not
              either running or a spare.')

We fixed this issue and took the opportunity to go through the code
and fix similar issues. We also removed the explicit conversion from
an object to a string when this was not necessary and improved the
aforementioned message:

'Server (%s) is either not running or is a spare.'