HAM-271: No error message appear if the add_shard (any FABRIC command) command is wrong (having wrong number of parameters).
The problem occurs in this case beccause a procedure having a wrong listing
of parameters gets treated as an asynchronous procedure.

for example in the the following command

mysqlfabric sharding add_shard 1 990 groupid-2 ENABLED 101

The synchronous input is computed using the following code

   @staticmethod
    def wait_for_procedures(procedure_param, synchronous):
        [snip function docs]....
        assert(len(procedure_param) == 1)
        synchronous = str(synchronous).upper() not in ("FALSE", "0")
        [snip rest of function]

The above command causes the synchronous = 101

synchronous = str(synchronous).upper() not in ("TRUE", "1")
synchronous = 101 in ("TRUE","1")
synchronous = False

Hence the command is executed as an asynchronous command.

When an asnchronous command is executed detailed status of the command execution
is not printed since it is unsafe to access the procedure status during the
execution of an ansynchornous procedure. Hence only the UUID is executed.

This causes the command output to contain only the uuid generated for the
procedure execcution.

A workaround for this is to change the code into the following

synchronous = str(synchronous).upper() not in ("FALSE", "0")

Thus in this case unless synchronous is explicitly set of FALSE, the value is
retailed as True. The user will thus get complete execution status and there is
a chance that the user will not turn off synchronous execution by mistake.

This is also keeping in with the principle with which the code itself was
designed,

for example,

ADD_SHARD = _events.Event("ADD_SHARD")
class AddShard(ProcedureShard):
    """Add a shard.
    """
    group_name = "sharding"
    command_name = "add_shard"
    def execute(self, shard_mapping_id, group_id, state="DISABLED",
                lower_bound="None", synchronous=True):
                
Note that in the above code snippet synchronous is True by default. Thus the fix
keeps with the above philosophy.