HAM-316: Configuration file should be in /etc/mysql.
Configuration files for all tools should be placed in
the common configuration file directory, which on Linux
is /etc/mysql under the installation prefix.

This patch move the default installation location from
/etc/fabric/main.cfg to /etc/mysql/fabric.cfg.

It also adds some documentation on the mysqlfabric
utility mainly to provide information to the user on
where the configuration file is installed.