HAM-283 Automatic switchover fails if group activation is done before adding slave
HAM-283 was causing the slave to not take over from the master,
once the master crashes. This was happening, as a result of the
sharding setup, that was trying to stop the replication on a master
that does not exist anymore.

The patch takes care to not setup(and stop) replication with masters
that do not exist.