HAM-364: test_promote (test_replication_events.py) is failing
test_mysql_server.py creates a MySQLServer object with a fake UUID
and tries to bind it to a valid MySQL instance. The connection is
created but the binding does not happen because the UUIDs do not
match. However this invalid connection was returned to a pool when
the object was destroyed.

Other test cases (e.g. test_promote) created MySQLServer objects
with the same fake UUID leading to unexpected failures.

To fix the problem, we purge the connection from the pool.