HAM-186 A server must belong to a single group
This patch defines a 1-N relationship between groups and servers.
Before the patch we had the assumption that a server could belong
to more than one group as we had thought that this was necessary
to implement replication among groups in the global updates.

The following functions were changed:

. Group.add_server()
. Group.remove_server()
. Group.servers()

The following function was eliminated:

. Group.group_from_server()

The import test and function had to be updated because they were
considering that a server could belong to more than group.

We also took the opportunity to remove some "\" that were not
needed and changed some functions that were returning a set and
made them return a list which is faster.