BUG#17639666: BUG#17655819: BUG#70694: MYSQLFABRIC IS NOT EXECUTABLE ON WINDOWS
On windows, "mysqlfabric" is copied to a script folder but cannot be 
directly executed because an executable file must have an extension
such as  *.exe, *.bat or *.py. 

This problem obliges users to use fabric as follows: 

c:\python2x\python c:\python2x\scripts\mysqlfabric 

This may be a little bit inconvenient and is not very intuitive. To fix
that, we copy "mysqlfabric.py" when windows is used but keep "mysqlfabric"
when other systems are used.