BUG#17804807:BUG#70924 SERVER OBJECT DOES NOT HAVE A CLEAN DESIGN TO REPRESENT SCALING OUT SERVERS
There is no information in the server that specifies its likelihood of
accepting read-only transactions. This along with the fact that the
field that stores whether a server accepts read-only or read-write
operations is also used to store information on the server's
availability make it difficult to define scale-out servers in Fabric.

In this patch, we have used the status field to only identify the
server's availability and created two additional fields. The first
field, named mode, is used to identify the type of transaction,
read-only or read-write, a server could accept and the second, named
weight, would denote its likelihood of getting a request.

So after this patch, we have:

Server's status that might be: [FAULTY, SPARE, SECONDARY, PRIMARY]
Server's mode that might be: [OFFLINE, READ_ONLY, WRITE_ONLY, READ_WRITE]
Server's weight which is a float value greater than 0.0