WL#7401: Read only remote commands in Fabric should *NOT* pass through the executor
The commands that update the state store or change the
topology of Fabric are passed through Fabric's multi-threaded
executor that checks to see if there are any commands that
can be run in parallel to increase the throughput and sequences
the commands that conflict.

The read-only methods on the other hand, do not conflict
and need not pass through the executor that checks for conflicts.
There is also the problem that these might pose pseudo conflicts
with the executing commands when actually there are none. Hence the
read-only methods can be executed directly taking advantage of the ACID
implementation of the underlying state store (MySQL).

Methods that need to be changed

group
-----
lookup_servers
check_group_availability
lookup_groups

server
------
lookup_uuid

sharding
--------
lookup_servers
list_mappings
list_definitions
lookup_mappings