WL#7423: Use the Shard Mapping ID for GLOBAL server lookups.
Currently the "GLOBAL" server lookup happens using the name of the table
that is being sharded. But the problem with this method of lookups is
that it is difficult to find out from a GLOBAL server query which table
it belongs to.

The best way to go forward is to use the Shard Mapping ID (for GLOBAL lookups).