BUG#17592301: MYSQLFABRIC MANAGE START HANGS
Currently, Fabric has the following set of threads:

- MainThread which silently waits until someone executes "mysqlfabric
  manage stop"

- XML-RPC-Server which listens to incoming requests

- XML-RPC-Session(s) which are responsible for processing incoming
  requests. Read-only requests are directly processed by the
  XML-RPC-Session threads. However, write requests are delegated to
  an Executor thread which provides some concurrency mechanisms.

- Executor(s) which processes write requests

There were some places where exceptions were not caught and by
consequence the caller thread was killed. If the caller was the
XML-RPC-Server, no incoming request would be accepted. If the caller
was the XML-RPC-Session, no incoming request would be processed.
If the caller was the Executor, no write requests would be processed:

- The thread.start() was failing in some windows platforms and the
  exception was not caught thus killing the caller thread.

- Some exceptions (e.g. "too many connections") while accessing the
  backing store were not caught thus killing the caller thread.

We have also managed to find out that the XML-RPC-Session could block
waiting for a dead client and surprisingly for clients that finished
normally but left an open connection (BUG-17702315). This would make
"mysqlfabric manage stop" hang because the shutdown routine waits
until all the XML-RPC-Session threads finish.

Finally, there was a concurrency issue between the shutdown and the
scheduler. One could schedule a procedure while the shutdown was
taking place thus making the XML-RPC-Server wait for a procedure that
would never be executed.

To fix these issues, we have:

- started catching the aforementioned exceptions

- made the "mysqlfabric manage stop" not wait for the XML-RPC-Session

- fixed the concurrency issue between the shutdown and the scheduler

We also have taken the opportunity to improve the code that access the
backing store. Now it tries to re-connect to the backing store if there
is a problem with it (.e.g too many connections). This is done only
when a new transaction is started and never in the middle of a
transaction, the attempt to reconnect would hide problems and could
lead to inconsistencies otherwise.