BUG#17633546: CHECKPOINT ROUTINES LEAD TO DEADLOCKS
The statement used to register a procedure/job was the following:

INSERT INTO checkpoints(proc_uuid, lockable_objects, job_uuid,
sequence, action_fqn, param_args, param_kwargs) SELECT %s, %s, %s,
coalesce(MAX(sequence), 0) + 1, %s, %s, %s FROM checkpoints WHERE
proc_uuid = %

The job's sequence number was assigned while inserting the procedure/
job by selecting the max sequence number assigned and thus the "select"
 was leading to deadlocks among concurrent transactions that tried to
insert a procedure/job. To fix that, we have added an sequence
generator into the procedure class. With this approach, there is no
concurrency issues as a job belongs to a single procedure and one
single thread is responsible for executing a procedure.

In the context of this bug, we also took the opportunity to rename the
term schedule to register in the checkpoint module. This was done
because the checkpoint module does not schedule procedures/jobs but
only registers them.

There are still many things that must/can be improved within the
checkpoint module but we kept this patch simple and will do other
improvements in the context of other patches.