BUG 17852472 - FORMAT OF "GROUP.LOOKUP_SERVERS" HAS AN EXTRA STRUCT WRAPPING
The WL#7401 has changed the response format of read-only commands such as
the "group.lookup_servers" to include an extra struct wrapping the return
value. These commands return a dictionary as follows: 

{
 "success" : {False | True}
 "message" : Exceptions
 "return"  : Returned value
}

When called from the command-line the dictionary is wrapped and produces
the following output:

Command : 
{ return = 
  {
    "success" : {False | True}
    "message" : Exceptions
    "return"  : Return value
   }
} 

We think this extra indirection to represent the return value is not good
and is fixed in this patch. So the following output is produced when a
read-only command is executed from the command-line:

Command : 
{ "success" : {False | True}
  "return"  : Return value or Exception
} 

We also have changed the response format in order to make it more in-line
with the read-write and dump commands. Instead of returning a dictionary,
we now return a list/array as follows:

    if exception:
        return [False, traceback.format_exc(), True]
    else:
        return [True, "", status]