BUG#18124108: BUG#71428: FABRIC.SERVER.SET_STATUS() IS INCONSISTENT WITH FABRIC.STORE.DUMP_SERVERS()
The mysqlfabric.server.set_status function requires a uuid and a status
as parameters. The first parameter uniquely identifies the server and the
second determines its status which is one of the following strings:
"PRIMARY", "SECONDARY", "SPARE" or "FAULTY". On the other hand, the dump
interface uses an integer to represent a server's status.

Having the same information represented in different ways is not good
from an API's perspective. In this patch, we make make the set_status()
accept not only a string but also an integer as status.

So an integer is the common representation. However, we believe that the
string representation makes it easier to use the set_status() function
from the command-line. The string was not adopted as common representation
because this would would increase the dump's footprint.