BUG#71525: BUG#17702237: IS_CONNECTED() IS CHECKED EVERY TIME A STATEMENT IS EXECUTED
This patch contains four fixes that are related to routines that check whether
the server object is connected to a database server.

1 - Every time a statement was executed, Fabric was checking whether the
connection was valid or not before the execution. This is not necessary and we
only need to guarantee that the connection pool returns a valid connection.

So we removed the call to cnx.is_connected() from exec_mysql_stmt() and also
took the opportunity to remove a similar call from
reestablish_mysql_connection().

2 - Some calls to the is_alive() method were unnecessary because the connect()
method was always called before calling the is_alive() method. The connect()
method provides a valid connection to a server and executes some initialization
queries against the server as well guaranteeing that it is alive and kicking

So we removed the unnecessary calls. In some cases though, the is_alive() method
was replaced by a call to is_connected() method which verifies whether a valid
connection was retrieved by calling connect().

3 - The is_alive() method was supposed to be used by the failure detector
routine but was used in several places.

So we fixed that in the previous steps and changed the is_alive() method making
it try to connect to a server in order to check whether it is alive or not.

4 - The Group.servers() method was automatically establishing a connection to
all the servers in a group when this sometimes was not necessary.

Now we let the caller decide whether a connection to a server should be
established or not.