WL#7528 Adding support for optional arguments in Fabric
This worklog aims to add the ability define optional arguments for Fabric
commands.

The MySQLFabric command line is currently described as follows,

mysqlfabric <command-group> <command> <arguments>

E.g.

mysqlfabric group lookup_servers <group-id> <uuid> <status> <mode> <synchronous>

All the arguments that are passed to a mysqlfabric command are compulsory. If
any of them need to be skipped, currently it is done by passing a dummy place-
holder (empty string "") to them. This is especially true for the arguments that
occur in the middle of defined ones. For example, in the above command to search
for the "PRIMARY" in a group, we would do

mysqlfabric group lookup_servers <group-id> "" PRIMARY "" <synchronous>

Notice how the empty arguments are skipped using "".

However the above are optional arguments, and they should be skipped with a more
unambiguous syntax, for example in the above case, it should be possible to
optionally specify status using a syntax like the following,

mysqlfabric sharding lookup_servers <group-id> --status="PRIMARY"

Notice how the status is specified and the other optional arguments are skipped.