BUG#71448: BUG#18138545 - GROUP ADD FAILS IF ACCOUNT HAS NO ACCESS TO MYSQL SCHEMA
Fabric requires a user with all privileges granted on *.* in order to add a
server into a group. This is necessary because some provisioning operations
in the shard sub-system dump and restore all databases including mysql.

In this patch, we check the required permissions to add a server into a
group and create the appropriate infra-structure to test it.