Post-push fix for BUG#71512: BUG#18138545.
This patch fixes a documentation issue in the MySQLServer:has_privileges()
method. It also fixes some issues around the user/password used to connect
to servers from the tests. In the previous patch, the user/password was
created but the privileges were not flushed so some test cases were using
the root account.