BUG#71512: BUG#18153823: SERVERS IN THE SAME GROUP SHOULD USE A SINGLE USER/PASSWORD
Fabric requires to provide a user/password when a server is added to a group.
This makes the configuration process error prone as this user must also have
replication privileges at all servers that are member of the same group.

In order to ease the addition of a server into a group and avoid possible
configuration mistakes and errors, Fabric should require/assume a common user
for servers in a group or shard definition and when those are not defined, it
should use a global user/password.

In this patch, we introduce a global user/password and will consider the
other alternatives in the context of another bug or worklog. So after this
patch the following commands are affected:

- mysqlfabric group add "group" "host:port"
  . user/password was removed from the list of parameters

- mysqlfabric group import_topology "group-pattern" "description" "host:port"
  . user/password was removed from the list of parameters

- mysqlfabric server lookup_uuid "host:port"
  . user/password was removed from the list of parameters

We also took the opportunity to improve the ./lib/mysql/fabric/services/server.py
file and the returned values in the following functions:

- mysqlfabric group lookup_servers "group" --uuid=uuid
  . List with servers' info:
    [{"server_uuid" : uuid, "address" : address,
     "status" : status, "mode" : mode, "weight" : weight},
     {...}, ... 
    ]

- mysqlfabric group lookup_groups --group_id=group_id
  . List with groups' info:
    [{"group_id" : group_id, "description" : description,
     "master_uuid" : master_uuid, "failure_detector" : True/False},
     {...}, ...
    ]